stats: track packet check results per ISD-AS and DRKey protocol number
LF packet check results are now tracked for each ISD-AS and DRKey protocol that is listed as a peer in the configuration separately. The metrics are exposed on the telemetry path `lf/ia/stats`. Furthermore, `lf/ia/stats/list` provides a list of all combinations of ISD-AS and DRKey ptorocol numbers that are being tracked.